+++ packages/transaction-pay-controller/src/strategy/relay/relay-submit.test.ts
@@ -23,7 +23,14 @@ import {
   waitForTransactionConfirmed,
 } from '../../utils/transaction';
 
-jest.mock('../../utils/token');
+jest.mock('../../utils/token', () => ({
+  ...jest.createMockFromModule<typeof import('../../utils/token')>(
+    '../../utils/token',
+  ),
+  normalizeTokenAddressForMM:
+    jest.requireActual<typeof import('../../utils/token')>('../../utils/token')
+      .normalizeTokenAddressForMM,
+}));
 jest.mock('../../utils/transaction');
 jest.mock('../../utils/feature-flags');